+++ example/lib/home/home.dart
@@ -4,7 +4,6 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:universal_ble/universal_ble.dart';
-import 'package:universal_ble_example/data/capabilities.dart';
 import 'package:universal_ble_example/data/mock_universal_ble.dart';
 import 'package:universal_ble_example/home/widgets/scan_filter_widget.dart';
 import 'package:universal_ble_example/home/widgets/scanned_devices_placeholder_widget.dart';
@@ -148,7 +147,7 @@ class _MyAppState extends State<MyApp> {
                     });
                   },
                 ),
-                if (Capabilities.supportsBluetoothEnableApi &&
+                if (BleCapabilities.supportsBluetoothEnableApi &&
                     bleAvailabilityState == AvailabilityState.poweredOff)
                   PlatformButton(
                     text: 'Enable Bluetooth',
@@ -159,7 +158,7 @@ class _MyAppState extends State<MyApp> {
                       );
                     },
                   ),
-                if (Capabilities.requiresRuntimePermission)
+                if (BleCapabilities.requiresRuntimePermission)
                   PlatformButton(
                     text: 'Check Permissions',
                     onPressed: () async {
@@ -172,7 +171,7 @@ class _MyAppState extends State<MyApp> {
                       }
                     },
                   ),
-                if (Capabilities.supportsConnectedDevicesApi)
+                if (BleCapabilities.supportsConnectedDevicesApi)
                   PlatformButton(
                     text: 'Connected Devices',
                     onPressed: () async {