+++ lib/src/models/model_exports.dart
@@ -8,4 +8,5 @@ export 'package:universal_ble/src/models/availability_state.dart';
 export 'package:universal_ble/src/models/ble_connection_state.dart';
 export 'package:universal_ble/src/models/ble_device.dart';
 export 'package:universal_ble/src/models/ble_command.dart';
+export 'package:universal_ble/src/models/ble_capabilities.dart';